Tests: Added tests for the `index.js` file.
